svn path=/branches/work/kopete/iup-isi-group/; revision=861061

+++ protocols/jabber/libiris/iris/jabber/filetransfer.cpp
@@ -1,6 +1,6 @@
 /*
  * filetransfer.cpp - File Transfer
- * Copyright (C) 2004  Justin Karneges
+ * Copyright (C) 2004  Justin Karneges <justin@affinix.com>
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -592,7 +592,7 @@ bool JT_FT::take(const QDomElement &x)
 		if(!file.isNull()) {
 			QDomElement range = file.elementsByTagName("range").item(0).toElement();
 			if(!range.isNull()) {
-				int x;
+				qlonglong x;
 				bool ok;
 				if(range.hasAttribute("offset")) {
 					x = range.attribute("offset").toLongLong(&ok);
